오류 처리 완료.

+++ src/main/java/kr/hhplus/be/server/reservation/usecase/ReserveConcertUseCase.java
@@ -66,9 +66,8 @@ public class ReserveConcertUseCase {
 		reservation.setHoldExpiresAt(LocalDateTime.now().plusMinutes(HOLD_DURATION_MINUTES));
 
 		// 가격 계산 (센트 단위)
-		long priceInCents = seat.getConcertSchedule().getConcertPrice()
-				.multiply(new BigDecimal(100))
-				.longValue();
+		BigDecimal priceInCents = seat.getConcertSchedule().getConcertPrice()
+				.multiply(new BigDecimal(100));
 
 		reservation.setAmountCents(priceInCents);
